Fix #34: Migrasi monetary columns ke integer cents
- Migration: decimal(12,2) → bigInteger, konversi data ×100
- Models: casts decimal:2 → integer
- Form Requests: prepareForValidation() konversi input rupiah → cents
- Controllers: hapus (float) casts dan bccomp()
- Views: display / 100
- Tests/Factories: form submission rupiah, stored assertions cents
- Fix migration #23: drop index sebelum drop column room_name (SQLite compat)
- Fix pre-existing test failures akibat room_name column di-drop
- Add room_name validation to StoreSantriRequest & UpdateSantriRequest

+++ app/Modules/Saas/Controllers/SaasDashboardController.php
@@ -37,8 +37,8 @@ class SaasDashboardController extends Controller
                 'expired_tenants' => Tenant::query()->where('subscription_status', Tenant::SUBSCRIPTION_EXPIRED)->count(),
                 'deleting_tenants' => Tenant::query()->where('subscription_status', Tenant::SUBSCRIPTION_DELETING)->count(),
                 'platform_users' => User::query()->count(),
-                'total_revenue' => (float) TenantBillingNote::query()->sum('amount'),
-                'revenue_this_month' => (float) TenantBillingNote::query()
+                'total_revenue' => (int) TenantBillingNote::query()->sum('amount'),
+                'revenue_this_month' => (int) TenantBillingNote::query()
                     ->whereBetween('paid_at', [now()->startOfMonth(), now()->endOfMonth()])
                     ->sum('amount'),
             ],
@@ -90,7 +90,7 @@ class SaasDashboardController extends Controller
 
                 return [
                     'label' => $month->translatedFormat('M Y'),
-                    'amount' => (float) TenantBillingNote::query()
+                    'amount' => (int) TenantBillingNote::query()
                         ->whereBetween('paid_at', [$month->copy()->startOfMonth(), $month->copy()->endOfMonth()])
                         ->sum('amount'),
                 ];
